Rename BlockThreshold to HarmBlockThreshold
This better matches the backend API.
