+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/internal/util/conversions.kt
@@ -27,10 +27,8 @@ import com.google.firebase.vertexai.common.shared.FunctionCall
 import com.google.firebase.vertexai.common.shared.FunctionCallPart
 import com.google.firebase.vertexai.common.shared.FunctionResponse
 import com.google.firebase.vertexai.common.shared.FunctionResponsePart
-import com.google.firebase.vertexai.common.shared.HarmBlockThreshold
 import com.google.firebase.vertexai.type.BlobPart
 import com.google.firebase.vertexai.type.BlockReason
-import com.google.firebase.vertexai.type.BlockThreshold
 import com.google.firebase.vertexai.type.Candidate
 import com.google.firebase.vertexai.type.CitationMetadata
 import com.google.firebase.vertexai.type.Content
@@ -41,6 +39,7 @@ import com.google.firebase.vertexai.type.FunctionCallingConfig
 import com.google.firebase.vertexai.type.FunctionDeclaration
 import com.google.firebase.vertexai.type.GenerateContentResponse
 import com.google.firebase.vertexai.type.GenerationConfig
+import com.google.firebase.vertexai.type.HarmBlockThreshold
 import com.google.firebase.vertexai.type.HarmCategory
 import com.google.firebase.vertexai.type.HarmProbability
 import com.google.firebase.vertexai.type.HarmSeverity
@@ -141,13 +140,18 @@ internal fun ToolConfig.toInternal() =
     )
   )
 
-internal fun BlockThreshold.toInternal() =
+internal fun HarmBlockThreshold.toInternal() =
   when (this) {
-    BlockThreshold.NONE -> HarmBlockThreshold.BLOCK_NONE
-    BlockThreshold.ONLY_HIGH -> HarmBlockThreshold.BLOCK_ONLY_HIGH
-    BlockThreshold.MEDIUM_AND_ABOVE -> HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE
-    BlockThreshold.LOW_AND_ABOVE -> HarmBlockThreshold.BLOCK_LOW_AND_ABOVE
-    BlockThreshold.UNSPECIFIED -> HarmBlockThreshold.UNSPECIFIED
+    HarmBlockThreshold.NONE ->
+      com.google.firebase.vertexai.common.shared.HarmBlockThreshold.BLOCK_NONE
+    HarmBlockThreshold.ONLY_HIGH ->
+      com.google.firebase.vertexai.common.shared.HarmBlockThreshold.BLOCK_ONLY_HIGH
+    HarmBlockThreshold.MEDIUM_AND_ABOVE ->
+      com.google.firebase.vertexai.common.shared.HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE
+    HarmBlockThreshold.LOW_AND_ABOVE ->
+      com.google.firebase.vertexai.common.shared.HarmBlockThreshold.BLOCK_LOW_AND_ABOVE
+    HarmBlockThreshold.UNSPECIFIED ->
+      com.google.firebase.vertexai.common.shared.HarmBlockThreshold.UNSPECIFIED
   }
 
 internal fun Tool.toInternal() =
