+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/HarmBlockThreshold.kt
@@ -19,7 +19,7 @@ package com.google.firebase.vertexai.type
 /**
  * Represents the threshold for some [HarmCategory] that is allowed and blocked by [SafetySetting].
  */
-enum class BlockThreshold {
+enum class HarmBlockThreshold {
   /** The threshold was not specified. */
   UNSPECIFIED,
